Begin implementing input manager
Former-commit-id: 48a25bb69dad16d32479972d7bd649c2f9289ede

+++ Assets/Scripts/Util/IntRange.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using UnityEngine;
 
 [Serializable]
 public class IntRange {
@@ -8,14 +9,15 @@ public class IntRange {
     public int min;
     public int max;
 
-    public IntRange(int _min, int _max)
-    {
+    public IntRange(int _min, int _max) {
         min = _min;
         max = _max;
     }
 
-    public float Random
-    {
-        get { return UnityEngine.Random.Range(min, max+1); }
+    // Seems to be broken :( wasn't important enough to fix yet
+    public int Random {
+        get {
+            return UnityEngine.Random.Range(min, max + 1);
+        }
     }
 }
